Blob/File constructors/slice method shouldn't throw on invalid types
Per spec[1] and tests[2] the constructors for Blob and File and
their slice() methods should not throw if non-printable-ASCII type
is supplied; instead it should use "" as the type. In addition, the
type should be consistently lower-cased, by both the constructors
and slice() method.

[1] https://w3c.github.io/FileAPI/
[2] https://github.com/w3c/web-platform-tests/tree/master/FileAPI

BUG=509786
R=kinuko@chromium.org

Review-Url: https://codereview.chromium.org/1362963003
Cr-Original-Commit-Position: refs/heads/master@{#455965}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0b46f3a24bf36d7cf91ec2d88132d15563b85d0e