DevTools: add more tests to fixate bindings behavior
NetworkProject's cleanup logic currently relies on resources in
resourceTreeFrame. This patch adds tests with scripts and
stylesheets which don't have underlying resource:
- contentscript-* tests validate bindings creating and removing
  UISourceCodes for content scripts
- dynamic-* tests validate bindings creating and removing
  UISourceCodes for dynamically-created styles

As a drive-by, the patch improves the InspectorTest.dumpWorkspace()
method to also produce a diff with the previous workspace dump.
This helps to read expectations a lot.

R=dgozman
BUG=670180

Review-Url: https://codereview.chromium.org/2738033002
Cr-Original-Commit-Position: refs/heads/master@{#455973}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 074f0c4a4ad0b9cf594780b34bde720f57a177b8