Add registered custom property wrapping support to CSSAnimatableValueFactory
This patch enables AnimatableValues to be created from registered
custom property CSSValues. This does not enable interpolation of
the registered custom properties via AnimatableValues, this is only
required because AnimatableValues are still used by CSSTransitions
as a way of testing equality between animation keyframes and
computed style values.

This patch is a refactor and does not change behaviour.

BUG=671904

Review-Url: https://codereview.chromium.org/2723883005
Cr-Original-Commit-Position: refs/heads/master@{#456023}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3cfa3f891108a0b22cdf04122cbde1f85d9e87e7