Implement BackgroundFetchManager::{get, fetch}()
This CL implements the get() and fetch() methods for Background Fetch.
The BackgroundFetchRegistration interface and BackgroundFetchOptions
and IconDefinition dictionary types are added too.

The method has not been hooked up with the browser process yet, as the
Mojo interface has not been implemented yet.

Tests are provided, and will be moved to the WPT directory once it has
been approved upstream:
  https://github.com/w3c/web-platform-tests/pull/5094

BUG=692534

Review-Url: https://codereview.chromium.org/2745543002
Cr-Original-Commit-Position: refs/heads/master@{#456144}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6e65af76a14bb2230117cf4bc7dfc76d5a4ed24b