Avoid calling isMainThread() in the common cases when checking ScriptForbiddenScope::isScriptForbidden().
Usually when s_scriptForbiddenCount is true we're about to crash, so check that
first then check which thread it is. This means a background thread will read
the potentially racy value of the counter, but it'll always return false when
it checks isMainThread() next.

Review-Url: https://codereview.chromium.org/2709763003
Cr-Original-Commit-Position: refs/heads/master@{#456284}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3b4b07ced6315a25960ca5a51dd1b7ba88de53db