Implement the events for Background Fetch
This CL implements the events for Background Fetch. There are four:

  - BackgroundFetchEvent (onbackgroundfetchabort; base event)
  - BackgroundFetchClickEvent (onbackgroundfetchclick)
  - BackgroundFetchedEvent (onbackgroundfetched)
  - BackgroundFetchFailEvent (onbackgroundfetchfail)

Not all attributes are supported yet. Notably, the sequences of requests
and responses associated with a fetch will be exposed in a subsequent CL.

Tests are included, and are in addition to the IDL harness test in WPT.
I'm adding tests for the `fail` and `fetched` events in the next CL,
where they get Request and Response knowledge.

BUG=692534

Review-Url: https://codereview.chromium.org/2745573002
Cr-Original-Commit-Position: refs/heads/master@{#456296}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4c13484884ae6388d116d6d9d84c9fc42479630d