Implements CSSPropertyAPI for the fragmentation element properties.
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyWidows, CSSPropertyOrphans and
CSSPropertyWebkitFragmentationOrdinalGroup from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
border radius properties is stored in a CSSPropertyDescriptor, and is
called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIFragmentationGroup.cpp to the BUILD.gn file.
- Adds api_class=CSSPropertyAPIFragmentationGroup flag to each fragmentation class
  property in CSSProperties.in, which indicates that
  CSSPropertyAPIFragmentationGroup.h is generated.
- Moves the parsing logic for fragmentation properties from CSSPropertyParser.cpp
  to CSSPropertyAPIFragmentationGroup.cpp, which implements CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2639303004
Cr-Original-Commit-Position: refs/heads/master@{#456320}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b5e82dce09c880f7aa92a7a2203dc4b0e64c0e84