Include CSS property aliases when generating CSSPropertyMetadata
This change updates the metadata of alias properties to match the
property being aliased. This also adds the alias properties to
the generated CSSPropertyMetadata. Previously they were omitted and
any query on them would return false unconditionally.

This change also makes the ordering of properties in
CSSPropertyMetadata deterministic instead of Python's hash order.
This is a side effect of using self._properties_including_aliases
instead of self._properties where the former is a list and the latter
is a dict.

This is a refactor patch that does not change behaviour.
The inclusion of aliases in CSSPropertyMetadata is required for
future modifications to it.

BUG=699109

Review-Url: https://codereview.chromium.org/2746453004
Cr-Original-Commit-Position: refs/heads/master@{#456330}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ac041cfab01f509ef22e0f6fedbad3b421bbe20a