Disable NGTextLayoutAlgorithmTest.TextFloatsAroundFloatsBefore on Android
https://codereview.chromium.org/2739683006 introduced a test failure on
Android:
https://build.chromium.org/p/chromium.webkit/builders/WebKit%20Android%20%28Nexus4%29/builds/62131.

I would ordinarily revert the CL but it doesn't revert cleanly as
https://codereview.chromium.org/2735703005 landed afterward, and I don't have
enough domain expertise to know whether these two patches are related or
independent. As a result I'm just disabling the test on Android. This however
means that the bug is still present in the tree, so crbug.com/700868 needs
to be fixed ASAP.

BUG=700868
TBR=glebl

Review-Url: https://codereview.chromium.org/2748683002
Cr-Original-Commit-Position: refs/heads/master@{#456361}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2ac5618843f8e7b3044b9931c571d45aaf2915ca