Implement the Request/Response bits of Background Fetch
This CL implements the parts of the Background Fetch API that deal with
requests and responses, notably in the events. Following this, the Web
Exposed API that we're aiming for in MVP is implemented.

Tests for the `backgroundfetched` and `backgroundfetchfail` events have
been added, pending upstreaming to WPT like the others.

BUG=692534

Review-Url: https://codereview.chromium.org/2744873002
Cr-Original-Commit-Position: refs/heads/master@{#456383}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f0a182b5ebc9c86a6dd3df10cb0ebdbfb73c43dd