Replace descriptor_only with is_property and is_descriptor in CSSProperties.json5
This change adds the ability to query whether a CSSPropertyID is the name
of a CSS descriptor. This is intended to be used by DCHECKs in future patches.

The flag is_descriptor didn't fit well with the existing descriptor_only flag.
As a solution to the naming it was decided that having is_property and
is_descriptor flags would be nicer.

BUG=699109

Review-Url: https://codereview.chromium.org/2743503002
Cr-Original-Commit-Position: refs/heads/master@{#456634}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ecf42a93545affe9be77a837cad4212ad69cd38c