[LayoutNG] Manage needsLayout() for inline
This patch changes NGBlockNode::CopyFragmentDataToLayoutBox() not to
clearNeedsLayout() for all inline children. It was done to run unit
tests when inline was still at an early stage. But at this point, it
causes multiple layouts or layouts not to run for some nodes.

This change revealed some cases where CollectInlines() does not collect
all inlines. To fix them, NGInlineNode constructor was changed to take
the parent LayoutBlowFlow.

As a side effect, NGInlineNode::NextSibling() was implemented, so that
NGBlockChildIterator can traverse correctly, except when continuation is
needed.

When to clearNeedsLayout() for inline still needs more thoughts and
experiments.

BUG=636993

Review-Url: https://codereview.chromium.org/2750673002
Cr-Original-Commit-Position: refs/heads/master@{#456653}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 312584395a85fda56890f6174bf44b057f1d7417