[css-grid] Fix crash removing positioned grid item
When we add or remove a positioned item we don't need to mark
the grid as dirty, because positioned items do not affect the layout
of the grid at all.

This was causing a crash when a positioned item was removed
after a layout. As after the positioned item was removed,
the method LayoutGrid::layoutBlock() was not called,
so when the grid was repainted we got a crash.

BUG=700383
TEST=fast/css-grid-layout/grid-crash-remove-positioned-item.html

Review-Url: https://codereview.chromium.org/2748983003
Cr-Original-Commit-Position: refs/heads/master@{#457061}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eb7634d5a91a495af9495164ffb2e5f1f91f7e09