[LayoutNG] Refactor NGLineBuilder for floats in inline context
This patch adds a placeholder for floats in inline context, in the
similar way as BreakingContext::handleFloat() does.

In doing so, the semantics of NGLineBuilder::SetEnd() was changed that:
1. Index is no longer needed, only end offset, since it needs to look at
   the item anyway.
2. Caller must incrementally call one item by one.

The second point isn't important, easy to take out if needed. It's
slightly simpler and that there until we see it's really needed.

BUG=636993

Review-Url: https://codereview.chromium.org/2749033004
Cr-Original-Commit-Position: refs/heads/master@{#457172}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 65a48689dd0b45f18b3fbd66e75b43b0b9c9964c