DevTools: finalize release note for v58
Per Pavel's feedback, I've changed the versioning logic for release notes to
be browser version-agnostic so we have the ability to show multiple release
notes for a given browser version.

I've also added in the links for v58 release note. Because we're planning on
launching it in Canary (currently v59) soon, I've taken off the version number
from the UI to avoid confusion but I can add it back on for the following release.

(This initial release is just to iron out any potential bugs in Canary).

BUG=682879

Review-Url: https://codereview.chromium.org/2751483002
Cr-Original-Commit-Position: refs/heads/master@{#457181}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38ff0f585179a4201aa1a3370995c30b357fd72d