bindings: Expand typedefs before generating union files.
Prior to this commit, the IDL compiler was failing to resolve typedefs in
unions before generating new union code. In practice, this meant an IDL
construct such as:

    typedef boolean Foo;
    void func((Foo or long) arg);

Generated FooOrLong.cpp and FooOrLong.h, and "Foo" was considered an IDL
interface instead of a simple bool. Since Foo.cpp and Foo.h do not exist,
and neither does the Foo class, this caused build failures.

We now resolve the typedefs earlier, and in the example above we would now
generate BooleanOrLong.cpp and BooleanOrLong.h, which works as expected.

Fixing this requires a few separate, intertwined changes:
* IdlUnionType.resolve_typedefs() was broken: the |typedefs| argument it
  receives is a dictionary whose keys are typedef names, not
  IdlTypeBase-derived objects, so |typedefs.get()| was always failing and
  returning the original, unresolved member objects.
* CodeGeneratorUnionType actually needs to call resolve_typedefs(). It was
  the only generator class not doing so. The mechanism needs to be different
  from the other classes though, as IdlUnionType is not a type defined in
  idl_definitions.py as is the case for the types handled by the other
  generators. Instead of using TypedefResolver, we simply call the relevant
  parts of it manually.
* The code manually adding union header includes to interfaces and partial
  interfaces has been dropped: there isn't enough information to resolve the
  typedefs at that point, which meant interface .cpp files would include
  FooOrLong.h, while interface .h files include BooleanOrLong.h. The test
  expectations IDL files show a reduction in unneeded includes, and the
  build seems to work fine.

BUG=701411
R=bashi@chromium.org,haraken@chromium.org,yukishiino@chromium.org

Review-Url: https://codereview.chromium.org/2749253002
Cr-Original-Commit-Position: refs/heads/master@{#457377}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 200763220d087ef3ccf735855398ec193db29658