Editing: Content of display:table elements should be editable.
There are no problems with adding arbitrary content to an element with
display:table if it's not <table>.

This CL fixes broken editing/deleting/display-table.html.

Because we don't want to change TextIterator behavior, this CL copies the old
implementation of isDispalyInsideTable() to TextIterator.cpp.

BUG=697315

Review-Url: https://codereview.chromium.org/2757553002
Cr-Original-Commit-Position: refs/heads/master@{#457393}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: af861c1f7d2bfe863febea4fd6a5d14483c30b48