Reduce runtime of osc-negative-freq test
Reduce the sample rate and duration of the osc-negative-freq test to
make it run faster for MSAN tests.  Tightened up the thresholds for
the new test.

On a Z620 machine running linux, the debug runtime was 1.414 sec. It
is now 0.860, roughtly 60% as long as before.

BUG=701418
TEST=osc-negative-freq.html

Review-Url: https://codereview.chromium.org/2751213002
Cr-Original-Commit-Position: refs/heads/master@{#457608}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: db917052bfb936546b2602cdb657c310778c6451