In wpt-update-expectations, ignore tests that unexpectedly pass.
In general, when the auto-importer is updating expectations and baselines, if a test unexpectedly passes, then it should be OK to just leave it alone and not update anything.

If the pass was a flaky pass, then the previous behavior would have been to try to rebaseline it, which would remove the expectation, which would later cause trouble if the test is flaky.

BUG=687442

Review-Url: https://codereview.chromium.org/2750273003
Cr-Original-Commit-Position: refs/heads/master@{#457647}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb5a41a23dda1f926f1d724650ad63d22bb823ac