node-set($x), when $x is empty, should not halt XSLT processing
node-set is polymorphic; if called with a string, it returns a
node-set containing a text node with that content. The libxslt roll in
r425586 broke this behavior because Blink's fork of the node-set
extension returned a text node without an owner document. Fix that.

BUG=689977

Review-Url: https://codereview.chromium.org/2750943004
Cr-Original-Commit-Position: refs/heads/master@{#457692}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3f6e482b7f13c1fc5bd68227321ef6da0ddf9937