Use CSSProperties.json5 format for nonproperties as well.
This patch modifies make_computed_style_base.py so that hardcoded
nonproperties use the same JSON format as CSSProperties.json5. The
list of dicts in NONPROPERTY_FIELDS has the same format and meaning
as the data in CSSProperties.json5.

This allows us to specify different templates for nonproperties.

This is prework for combining the code logic for properties and
nonproperties.

BUG=628043

Review-Url: https://codereview.chromium.org/2745023002
Cr-Original-Commit-Position: refs/heads/master@{#457694}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d56884370eecac968f54c5826958938d7c8f1a4c