Fix dir=auto for display: none subtrees.
The dir auto code was using layoutObject()->style() to find the direction()
style value, but this means for a display: none element it'd fail to find the
direction and then we'd get the wrong answer.

Instead we should use computedStyle(), then we get the cached style for
display: none elements (or things like <option>) and correctly invalidate
the style when the properties that impact the auto direction change.

Review-Url: https://codereview.chromium.org/2746363012
Cr-Original-Commit-Position: refs/heads/master@{#457703}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 204011460e76cb45440a4301e72f589df200257e