PeriodicWave returns sine wave generator if no coefficients given
If the PeriodWaveOptions option does not specify either real or imag
components, the returned PeriodicWave should result in a sine wave when
used with an OscillatorNode.

The spec was unclear and incorrect in saying that the optional
PeriodicWaveOptions dictionary had optional members specified, but actually
required at least one to be specified.  The spec was updated
to make this clear that everything is correctly optional, and
specifies what PeriodicWave should do PeriodicWaveOptions is given.

Tests updated for this and an additional test to verify that the
output is a sine wave.

BUG=695513
TEST=constructor/periodicwave.html

Review-Url: https://codereview.chromium.org/2712933003
Cr-Original-Commit-Position: refs/heads/master@{#457824}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 10c77a53e77e2ae475a00cbddb5da81f0f2965c4